Update popup blocked dialog
