+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "app-sdk",
-	"version": "1.3.0",
+	"version": "1.3.1",
 	"repository": "https://github.com/base/account-sdk",
 	"author": "Base",
 	"license": "MIT",
