+++ packages/app-sdk/package.json
@@ -1,6 +1,6 @@
 {
 	"name": "@startale/app-sdk",
-	"version": "1.3.0",
+	"version": "1.3.1",
 	"description": "Superapp SDK",
 	"keywords": [
 		"startale",
