+++ packages/app-sdk/src/ui/assets/StartaleLogo.tsx
@@ -1,0 +1,14 @@
+export const StartaleLogo = () => {
+
+	return (
+		<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="none" viewBox="0 0 60 60">
+			<g clip-path="url(#a)">
+				<path fill="#000" d="M48 0H12C5.373 0 0 5.373 0 12v36c0 6.627 5.373 12 12 12h36c6.627 0 12-5.373 12-12V12c0-6.627-5.373-12-12-12"/>
+				<path fill="#fff" d="M37.022 46.957 30 53.978l-7.02-7.02L30 39.935zM22.98 46.93h-9.93V37h9.93zm23.952 0h-9.93V37h9.93zM20.077 29.965l-7.02 7.021-7.022-7.021 7.022-7.02zm33.888 0-7.02 7.021-7.022-7.021 7.021-7.02zM22.98 22.93h-9.93V13h9.93zm14.022-9.928V13h9.93v9.93h-9.93v-9.89L30 20.042l-7.02-7.02L30 6z"/>
+			</g>
+			<defs>
+				<clipPath id="a"><path fill="#fff" d="M0 0h60v60H0z"/></clipPath>
+			</defs>
+		</svg>
+	)
+}
