+++ packages/app-sdk/src/ui/assets/index.ts
@@ -1,5 +1,6 @@
-export * from './BaseLogo.js'
 export * from './BasePayLogo.js'
 export * from './colors.js'
 export * from './fontFaceCSS.js'
 export * from './injectFontStyle.js'
+export * from './StartaleLogo.js'
+
